Moving all excel related code to a separate file
